Assignment3/requirements.txt
psycopg2==2.9.1
Flask==2.0.1
PyJWT==2.3.0
